Use babel's built-in option manager for loading configs
