+++ .babelrc
@@ -1,0 +1,5 @@
+{
+  "plugins": [
+    ["transform-object-rest-spread", { "useBuiltIns": true }]
+  ]
+}
