+++ .gitignore
@@ -3,4 +3,4 @@ node_modules/
 npm-debug.log
 
 coverage/
-.nyc_output
+lib/
