+++ src/index.js
@@ -2,49 +2,41 @@
 
 const path = require('path');
 const resolve = require('resolve');
+const pkgUp = require('pkg-up');
+const targetPlugin = require('babel-plugin-module-resolver').default;
 const mapModule = require('babel-plugin-module-resolver').mapModule;
-const findBabelConfig = require('find-babel-config'); // eslint-disable-line
+const OptionManager = require('babel-core').OptionManager;
 
-function findModuleAliasConfig(conf) {
-    if (conf.plugins) {
-        return conf.plugins.find(p => p[0] === 'module-resolver' || p[0] === 'babel-plugin-module-resolver');
+function getPlugins(file, target) {
+    try {
+        const manager = new OptionManager();
+        const result = manager.init({
+            babelrc: true,
+            filename: file,
+        });
+
+        return result.plugins.filter((plugin) => {
+            const plug = OptionManager.memoisedPlugins.find(item =>
+                item.plugin === plugin[0],
+            );
+
+            return plug && plug.container === target;
+        });
+    } catch (err) {
+        // This error should only occur if something goes wrong with babel's
+        // internals. Dump it to console so people know what's going on,
+        // elsewise the error will simply be squelched in the calling code.
+        console.error('[eslint-import-resolver-babel-module]', err);
+        console.error('See: https://github.com/tleunen/eslint-import-resolver-babel-module/pull/34');
+        return [];
     }
-    return null;
-}
-
-function getPluginOpts(config) {
-    const env = process.env.BABEL_ENV || process.env.NODE_ENV || 'development';
-
-    if (config) {
-        const pluginConfig = findModuleAliasConfig(config);
-
-        if (config.env && config.env[env]) {
-            const envPluginConfig = findModuleAliasConfig(config.env[env]);
-            if (envPluginConfig) {
-                if (pluginConfig) {
-                    return {
-                        root: [].concat(pluginConfig[1].root, envPluginConfig[1].root),
-                        alias: Object.assign({}, pluginConfig[1].alias, envPluginConfig[1].alias),
-                    };
-                }
-                return envPluginConfig[1];
-            }
-        }
-
-        if (pluginConfig) {
-            return pluginConfig[1];
-        }
-    }
-
-    return {};
 }
 
 function opts(file, config) {
-    return Object.assign(
-        {},
-        config,
-        { basedir: path.dirname(file) },
-    );
+    return {
+        ...config,
+        basedir: path.dirname(file),
+    };
 }
 
 exports.interfaceVersion = 2;
@@ -61,20 +53,18 @@ exports.interfaceVersion = 2;
 exports.resolve = (source, file, options) => {
     if (resolve.isCore(source)) return { found: true, path: null };
 
-    const babelConfig = findBabelConfig.sync(path.dirname(file));
-    const babelrcPath = babelConfig.file;
-    const config = babelConfig.config;
-    let cwd = babelrcPath
-        ? path.dirname(babelrcPath)
-        : process.cwd();
+    const projectRootDir = path.dirname(pkgUp.sync(file));
 
     try {
-        const pluginOpts = getPluginOpts(config);
-        if (pluginOpts.cwd !== 'babelrc') {
-            cwd = pluginOpts.cwd || cwd;
-        }
+        const instances = getPlugins(file, targetPlugin);
 
-        const src = mapModule(source, file, pluginOpts, cwd) || source;
+        const pluginOpts = instances.reduce((config, plugin) => ({
+            cwd: plugin[1] && plugin[1].cwd ? plugin[1].cwd : config.cwd,
+            root: config.root.concat(plugin[1] && plugin[1].root ? plugin[1].root : []),
+            alias: Object.assign(config.alias, plugin[1] ? plugin[1].alias : {}),
+        }), { root: [], alias: {}, cwd: projectRootDir });
+
+        const src = mapModule(source, file, pluginOpts, path.resolve(pluginOpts.cwd)) || source;
         return {
             found: true,
             path: resolve.sync(src, opts(file, options)),
