--- test/examples/components/sub/envonly/.babelrc
@@ -1,13 +1,0 @@
-{
-  "env": {
-    "test": {
-      "plugins": [
-        ["module-resolver", {
-          "alias": {
-            "subsub": "../sub"
-          }
-        }]
-      ]
-    }
-  }
-}
